feat/fix: Make some database improvements and getWeather reuseble
Make the getWheatherObject to its own function
Add a way to fill the database from a SQL file
Fix AdminTable table in the schema for the database
Make README a little more clear

Refs: #126, 128, 133
