+++ Backend/prisma/fill.sql
@@ -1,0 +1,4 @@
+INSERT INTO admintable
+(id, StartDate, DayAmount, DBDayAmount, TimeSets)
+VALUES
+(1, 'today', 3, 7, ARRAY['06:00-08:00', '11:00-12:00', '13:00-15:00', '15:00-17:00'])
